requirements.txt
streamlit==1.36.0
langchain-groq==0.1.5
sentence-transformers==2.7.0
numpy==1.26.4
python-dotenv==1.0.1
faiss-cpu==1.7.4
tqdm==4.66.5
